+++ components/ConfirmedCasesCard.vue
