+++ components/cards/ConfirmedCasesDetailsCard.vue
@@ -1,27 +1,27 @@
 <template>
   <v-col cols="12" md="6" class="DataCard">
-    <svg-card
+    <confirmed-cases-card
       :title="$t('検査陽性者の状況')"
       :title-id="'details-of-confirmed-cases'"
-      :date="Data.inspections_summary.date"
+      :date="Data.main_summary.date"
     >
       <confirmed-cases-table
         :aria-label="$t('検査陽性者の状況')"
         v-bind="confirmedCases"
       />
-    </svg-card>
+    </confirmed-cases-card>
   </v-col>
 </template>
 
 <script>
 import Data from '@/data/data.json'
 import formatConfirmedCases from '@/utils/formatConfirmedCases'
-import SvgCard from '@/components/SvgCard.vue'
+import ConfirmedCasesCard from '@/components/ConfirmedCasesCard.vue'
 import ConfirmedCasesTable from '@/components/ConfirmedCasesTable.vue'
 
 export default {
   components: {
-    SvgCard,
+    ConfirmedCasesCard,
     ConfirmedCasesTable
   },
   data() {
