+++ pages/index.vue
@@ -29,7 +29,7 @@
       />
     </div>
     <v-row class="DataBlock">
-      <tested-cases-details-card />
+      <confirmed-cases-details-card />
       <confirmed-cases-number-card />
       <confirmed-cases-attributes-card />
       <tested-number-card />
@@ -44,7 +44,7 @@ import { MetaInfo } from 'vue-meta'
 import PageHeader from '@/components/PageHeader.vue'
 import Data from '@/data/data.json'
 import News from '@/data/news.json'
-import TestedCasesDetailsCard from '@/components/cards/TestedCasesDetailsCard.vue'
+import ConfirmedCasesDetailsCard from '@/components/cards/ConfirmedCasesDetailsCard.vue'
 import ConfirmedCasesNumberCard from '@/components/cards/ConfirmedCasesNumberCard.vue'
 import ConfirmedCasesAttributesCard from '@/components/cards/ConfirmedCasesAttributesCard.vue'
 import TestedNumberCard from '@/components/cards/TestedNumberCard.vue'
@@ -53,7 +53,7 @@ import AgeCard from '@/components/cards/AgeCard.vue'
 export default Vue.extend({
   components: {
     PageHeader,
-    TestedCasesDetailsCard,
+    ConfirmedCasesDetailsCard,
     ConfirmedCasesNumberCard,
     ConfirmedCasesAttributesCard,
     TestedNumberCard,
